lab-05/test/requirements.txt
PyQt5==5.15.9
cryptography==41.0.1
Flask==2.0.1
Werkzeug==2.0.3
Pillow==10.0.0
